Huge change to the TrafficMap class: convert to prototype based methods.
In theory this will save memory and posibly run faster but frankly right now
we're not using the TrafficMap class in a way that's really conducive to
take advatage of that.

What we lose is the ability to do useful private members so we wind up exposing
properties like the live_cam list and such that while it doesn't hurt to be
exposed, it feels icky.

So basically it's just that it's the "righter" way to do it and who knows, if
we ever get to the point where we're displaying multiple traffic maps, then
it's a win.